% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copreg_bayes.R
\name{aux1}
\alias{aux1}
\title{Bayesian Gaussian Copula Regression}
\usage{
aux1(h, param, W1, data)
}
\description{
Internal helpers supporting the exported \code{CopRegBayes} sampler.
}
\keyword{internal}
