% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copreg_bayes.R
\name{print.endog_copula_bayes}
\alias{print.endog_copula_bayes}
\title{Print a Bayesian Copula Regression Fit}
\usage{
\method{print}{endog_copula_bayes}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{endog_copula_bayes}.}

\item{digits}{Number of significant digits to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Bayesian Copula Regression Fit
}
\seealso{
\code{\link[=CopRegBayes]{CopRegBayes()}}, \code{\link[=summary.endog_copula_bayes]{summary.endog_copula_bayes()}}
}
