% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copreg_bayes.R
\name{print.summary.endog_copula_bayes}
\alias{print.summary.endog_copula_bayes}
\title{Print a Summary of a Bayesian Copula Regression Fit}
\usage{
\method{print}{summary.endog_copula_bayes}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{summary.endog_copula_bayes}.}

\item{digits}{Number of significant digits to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Summary of a Bayesian Copula Regression Fit
}
